Intermediate: Add new algorithm for CISD wave function
The main code for the new algorith has been implemented
and it runs, but WRONGLY! (that is, it still needs some
debugging)

The CISD wave function seems to be OK (it passes in a rough
comparison with intN wave function). This also has to be rechecked!!

The code runs for Li2, but clearly not correctly.