Add comparison between general and CISD algorithms. Fix bugs
An extra option can be used, to calculate the matrices X and C,
using both algorithms, and compare them.

Fixes some bugs in the CISD algorithm, as well in the CISD and
int. normalised functions.

Code seems to be running properly for all cases tested.

Not sure if some signs in the Int. Norm. function are correct,
but it agrees with the part of the wave function implemented in CISD_WF