Fixes a bug (transforming int N wf to CISD wf)
The Csd for different irreps was incorrect.
(same calculation for different point groups were giving
different results)

This fixes this, however, now --algorithm orb_rotations
does not give the same result of --algorithm CISD anymore.
This implies that the same coefficient is incorrect when reading
the int norm. WF into a FCI wave function. This has to be
investigated and fixed