Uses max coincidence order for FCI when importing from IntermNorm
This adds the option to save the coefficients for determinants
with the orbitals in the order of maximum coincidence with the
reference. Adds functions to change between this and the ascending
order.
Note, however, that the string graphs are still in the reverse
lexical order as for orbitals in ascending order. This only changes
the sign of the coefficients

This also modifies the FCIWaveFunction.get_coefficients_from_int_norm_wf
to save the coefficients in the new way. Tests for this are OK.