Fix some bugs in analitycal and numerical Jac for mindist CC
Numerical and Analytical Jacobians match for several tests.
Numerical Jacobian is carried out in the maximum coincidence order

* Known issues:

CCD tests for Jacobian fail
   (commented in tests/implementation/test_cc_manifold.py)

Analytical Jacobian is still made with the convention of ordered
orbitals. This can be simplified to use maximum coincidence order, we
just have to ignore the sign returned by exc_on_strin.
This should be the next improvement