Fix bug for Hessian in min dist to CC manifold
The diagonal element of the hessian associated to the alpha,alpha
excitations of the same orbitals (that actually do not participate in
the cluster operator as it annihilates the reference, but we store a
zero amplitude for it) was zero, but this leads to a singular matrix.
We put 1.0 in these diagonal elements.

The tests of test_wf_cc_dist_to_fci that had error due to the singular
matrix now converge, to zero distance, as it should, but the
amplitudes are wrong (see below)

This commits also adds the calculation of the numerical Hessian, but
it is still buggy (does not pass all tests).