Use ordered orbitals in to optimise minD in CC manifold
The optimization of the distance in the CC manifold is carried
out with the convention of ordered orbitals. The order of maximum
coincidence is used only when transforming the coupled cluster wave
function in a FCI like wave function and only internally: an option to
get_coefficients_from_int_norm, of FCIWaveFunction, has been added,
that set ordered orbitals before returning.

Furthermore, the alpha and beta signs to make the convesion between
ordering conventions are calculated once and stored.