Use only the independent amplitudes for Jac and Hess in min dist CC
We were having a singular Hessian matrix, that cannot be inverted.
The reason is that we store, for the alpha/alpha and beta/beta
ampitudes, both t_ij^ab and t_ij^ba = -t_ij^ab. This redundancy
means that the associated columns in the Hessian matrix will be
linearly independent (one is minus the other) and the matrix is
singular. Now we use only one of the amplitudes (a<b) to construct
the Jacobian and the Hessian. The compatibility with the array that
store the amplitudes is implemented (see in particuar
IntermNormWaveFunction.n_indep_ampl and
IntermNormWaveFunction.update_amplitudes).

Tests for min dist to CC run well.