Towards a cython pure wave functions - TRANSITION COMMIT
This commits transforms further wave function modules to cython
All modules in wave_functions are already in cython, and compiling.

All codes that are, or are to be, obsolete were removed. Thus,
some codes are still broken, because they depend on these modules.
The full conversion is to be made.

Tests run for interm_norm and fci.